+++ packages/undercut/src/push/operations/sort.js
@@ -1,11 +1,11 @@
-import { assertFunctor } from "../../utils/assert.js";
-import { numbers, reverse, strings } from "../../utils/compare.js";
+import { assert, assertFunctor } from "../../utils/assert.js";
+import { numbers, strings } from "../../utils/compare.js";
 import { closeObserver } from "../../utils/observer.js";
+import { asc, desc } from "../../utils/ordering.js";
 
-export function sort(comparator, isReverse = false) {
+export function sort(comparator, order = asc) {
 	assertFunctor(comparator, "comparator");
-
-	const actualComparator = isReverse ? reverse(comparator) : comparator;
+	assert(order === asc || order === desc, `"order" must be the asc or the desc function.`);
 
 	return function* (observer) {
 		const items = [];
@@ -21,8 +21,16 @@ export function sort(comparator, isReverse = false) {
 			observer.throw(e);
 		} finally {
 			if (success) {
-				for (const item of items.sort(actualComparator)) {
-					observer.next(item);
+				items.sort(comparator);
+
+				if (order === asc) {
+					for (const item of items) {
+						observer.next(item);
+					}
+				} else {
+					for (let index = items.length - 1; index >= 0; index--) {
+						observer.next(items[index]);
+					}
 				}
 			}
 
@@ -31,10 +39,10 @@ export function sort(comparator, isReverse = false) {
 	};
 }
 
-export function sortNumbers(isReverse = false) {
-	return sort(numbers, isReverse);
+export function sortNumbers(order = asc) {
+	return sort(numbers, order);
 }
 
-export function sortStrings(isReverse = false) {
-	return sort(strings, isReverse);
+export function sortStrings(order = asc) {
+	return sort(strings, order);
 }