+++ CHANGELOG.md
@@ -22,12 +22,16 @@ This release adds Push Lines and makes API more consistent in general.
 ### Changed
 
 - **[BREAKING]** `toValue` pull target now throws on 0 items (you've expected a value, but got none).
-- **[BREAKING]** Moved `reverseComparator` utility to the `compare` namespace under `reverse` name.
 - **[BREAKING]** Removed Iterable utilities from `utils` entry.
 - **[BREAKING]** Renamed back `createPullLine` into just `pullLine`.
 - **[BREAKING]** Renamed `isNil` utility to `isNullish` (better consistency with ES naming like "Nullish Coalescing Operator").
+- **[BREAKING]** Replaced `isReverse` parameter of the `sort` operation with `order` parameter accepting `asc/desc` functions.
 - **[BREAKING]** Replaced `tryCloseIterator` helper with `closeIterator`.
 - Optimized `interleave` operation.
+
+### Removed
+
+- **[BREAKING]** Removed `reverseComparator` utility.
 
 ## [0.3.0] - 2019-07-15